+++ src/pages/Game.js
@@ -6,6 +6,7 @@ import Question from '../components/Question';
 import Answer from '../components/Answer';
 import Timer from '../components/Timer';
 import { revealedAction } from '../actions/gameAction';
+import { timerActions } from '../actions/correctAnswer';
 import Header from '../components/Header';
 
 class Game extends React.Component {
@@ -13,15 +14,19 @@ class Game extends React.Component {
     super(props);
 
     this.state = {
+      time: 30,
       number: 0,
       results: props.request,
     };
 
     this.nextQuestion = this.nextQuestion.bind(this);
+    this.timer = this.timer.bind(this);
+    this.stateInterval = this.stateInterval.bind(this);
   }
 
   componentDidMount() {
     const { count, name, email, assertions } = this.props;
+    this.stateInterval();
     const player = {
       name,
       assertions,
@@ -29,6 +34,26 @@ class Game extends React.Component {
       gravatarEmail: email,
     };
     localStorage.setItem('state', JSON.stringify({ player }));
+  }
+
+  stateInterval() {
+    this.setState({ time: 30 });
+    const ONE_SECOND = 1000;
+    const interval = setInterval(this.timer, ONE_SECOND);
+    this.setState({ interval });
+  }
+
+  timer() {
+    const { dispatchRevealed, updateTimer, isRevealed } = this.props;
+    const { time, interval } = this.state;
+    if (time > 0 && !isRevealed) {
+      this.setState((prevState) => ({
+        time: prevState.time - 1,
+      }), () => updateTimer(time));
+    } else {
+      dispatchRevealed(true);
+      clearInterval(interval);
+    }
   }
 
   nextQuestion() {
@@ -41,12 +66,12 @@ class Game extends React.Component {
     if (number < (results.length - 1)) {
       this.setState((prevState) => ({
         number: prevState.number + 1,
-      }));
+      }), () => this.stateInterval());
     }
   }
 
   render() {
-    const { number, results } = this.state;
+    const { number, results, time } = this.state;
     return (
       <div>
         <Header />
@@ -58,7 +83,7 @@ class Game extends React.Component {
               results={ results }
               nextQuestion={ this.nextQuestion }
             />
-            <Timer results={ results } />
+            <Timer results={ results } time={ time } />
           </div>
         )}
       </div>
@@ -72,10 +97,12 @@ const mapStateToProps = (state) => ({
   name: state.player.name,
   email: state.player.gravatarEmail,
   request: state.apiReducer.request,
+  isRevealed: state.game.isRevealed,
 });
 
 const mapDispatchToProps = (dispatch) => ({
   dispatchRevealed: (payload) => dispatch(revealedAction(payload)),
+  updateTimer: (time) => dispatch(timerActions(time)),
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(Game);
@@ -85,6 +112,8 @@ Game.propTypes = {
   assertions: PropTypes.number.isRequired,
   name: PropTypes.string.isRequired,
   email: PropTypes.string.isRequired,
+  updateTimer: PropTypes.number.isRequired,
+  isRevealed: PropTypes.bool.isRequired,
   request: PropTypes.arrayOf(PropTypes.any).isRequired,
   history: PropTypes.objectOf(PropTypes.any).isRequired,
   dispatchRevealed: PropTypes.func.isRequired,
